Unify CC/Vintage failure handling in dependency resolution tests
