+++ platforms/software/resources-s3/src/integTest/groovy/org/gradle/integtests/resource/s3/maven/MavenS3RepoErrorsIntegrationTest.groovy
@@ -18,7 +18,6 @@
 package org.gradle.integtests.resource.s3.maven
 
 import org.gradle.api.credentials.AwsCredentials
-import org.gradle.integtests.fixtures.executer.GradleContextualExecuter
 import org.gradle.integtests.resource.s3.AbstractS3DependencyResolutionTest
 import org.gradle.integtests.resource.s3.fixtures.MavenS3Module
 
@@ -64,11 +63,7 @@ task retrieve(type: Sync) {
         then:
         fails 'retrieve'
         and:
-        failure.assertHasDescription(
-            (GradleContextualExecuter.configCache)
-            ? "Configuration cache state could not be cached:"
-            : "Execution failed for task ':retrieve'."
-        )
+        assertResolutionTaskFailed(":retrieve")
         failure.assertHasCause("Could not resolve all files for configuration ':compile'.")
                 .assertHasCause('Could not resolve org.gradle:test:1.85')
                 .assertHasCause("Could not get resource '${module.pom.uri}'.")
@@ -93,11 +88,7 @@ repositories {
         fails 'retrieve'
         then:
         //TODO would be good to have a reference of the wrong configured repository in the error message
-        failure.assertHasDescription(
-            (GradleContextualExecuter.configCache)
-                ? "Configuration cache state could not be cached:"
-                : "Execution failed for task ':retrieve'."
-        )
+        assertResolutionTaskFailed(":retrieve")
         failure.assertHasCause("Could not resolve all dependencies for configuration ':compile'.")
                 .assertHasCause("Credentials must be an instance of '${AwsCredentials.class.getName()}'.")
     }
@@ -115,11 +106,7 @@ repositories {
         when:
         fails 'retrieve'
         then:
-        failure.assertHasDescription(
-            (GradleContextualExecuter.configCache)
-                ? "Configuration cache state could not be cached:"
-                : "Execution failed for task ':retrieve'."
-        )
+        assertResolutionTaskFailed(":retrieve")
         failure.assertHasCause("Could not resolve all dependencies for configuration ':compile'.")
                 .assertHasCause("S3 resource should either specify AwsImAuthentication or provide some AwsCredentials.")
 
@@ -134,11 +121,7 @@ repositories {
         fails 'retrieve'
 
         and:
-        failure.assertHasDescription(
-            (GradleContextualExecuter.configCache)
-                ? "Configuration cache state could not be cached:"
-                : "Execution failed for task ':retrieve'."
-        )
+        assertResolutionTaskFailed(":retrieve")
         failure.assertHasCause("Could not resolve all files for configuration ':compile'.")
         failure.assertHasCause(
                 """Could not find org.gradle:test:1.85.
@@ -172,11 +155,7 @@ Required by:
         expect:
         fails 'retrieve'
         and:
-        failure.assertHasDescription(
-            (GradleContextualExecuter.configCache)
-                ? "Configuration cache state could not be cached:"
-                : "Execution failed for task ':retrieve'."
-        )
+        assertResolutionTaskFailed(":retrieve")
         failure.assertHasCause("Could not resolve all dependencies for configuration ':compile'.")
         failure.assertHasCause("Authentication scheme 'auth'(BasicAuthentication) is not supported by protocol 's3'")
     }
