+++ testing/internal-integ-testing/src/main/groovy/org/gradle/integtests/fixtures/AbstractDependencyResolutionTest.groovy
@@ -16,6 +16,7 @@
 
 package org.gradle.integtests.fixtures
 
+import org.gradle.integtests.fixtures.executer.GradleContextualExecuter
 import org.gradle.test.fixtures.ivy.IvyFileRepository
 import org.gradle.test.fixtures.maven.MavenFileRepository
 
@@ -31,5 +32,22 @@ abstract class AbstractDependencyResolutionTest extends AbstractIntegrationSpec 
 
     MavenFileRepository mavenRepo(String name = "repo") {
         return maven(name)
+    }
+
+    /**
+     * Asserts that the task that performs dependency resolution has failed.
+     * With Configuration Cache, the task is expected to fail at serialization time.
+     * In vintage mode, the task is expected to fail at execution time.
+     *
+     * @param taskSelector the qualified name of the task (including leading {@code :})
+     */
+    void assertResolutionTaskFailed(String taskSelector) {
+        if (GradleContextualExecuter.configCache) {
+            failureDescriptionContains("Configuration cache state could not be cached:")
+            failureDescriptionContains(taskSelector)
+        } else {
+            def description = "Execution failed for task '${taskSelector}'."
+            failure.assertHasDescription(description)
+        }
     }
 }
